Remove management programs: they're not worth it, just use the cli
It's too much of a pain to try to interface with both systemd/openvpn
and a high-level torrenting library. Instead we should just manually
start/stop the vpn and start/stop torrents. Use transmission-daemon and
either transmission-remote-cli to run from the cli or the web interface
and port forward 9091/whatever through vagrant. Then maybe tack on a
daemon that monitors the ip and if it changes then log +/ kill the vm?